add repo and title filters
